indexers: remove the index manager.
This removes the index manager and moves needed
helpers in manager.go to common.go.
